TVB-2724 Put back methods cancel_or_remove_burst and stop_burst into BurstService, as logically they belong here
